requirements.txt
autopep8==1.6.0
bootstrap-py==1.0.2
certifi==2021.10.8
Django==2.2.12
django-bootstrap-form==3.4
django-bootstrap4==21.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
pep8==1.7.1
Pillow==9.0.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
